requirements.txt
astroid==2.0.1
autopep8==1.3.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
dj-database-url==0.5.0
Django==2.0.6
django-common-helpers==0.9.1
django-cors-middleware==1.3.1
django-coverage==1.2.4
django-extensions==2.1.0
django-nose==1.4.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djmail==1.1.0
docopt==0.6.2
facebook-sdk==2.0.0
flake8==3.5.0
google-auth==1.5.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
psycopg2-binary==2.7.5
py3dns==3.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pylint==2.0.1
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
validate-email==1.3
wrapt==1.10.11
